initial 11

+++ section-1/section-1.h
@@ -18,7 +18,7 @@ double exercise007(int n);
 double exercise008(int n);
 unsigned long long exercise009(int n);
 unsigned long long exercise010(int x,int n);
-int exercise011(int n);
+unsigned long long exercise011(int n);
 int exercise012(int n);
 int exercise013(int n);
 int exercise014(int n);
